Changed 0.8+ install instructions to avoid deletion of app.php and db.php
I was finding that using Transmit (ftp client), overwriting the `anchor` folder would actually delete my `app.php` and `db.php` files (which is kind of like "undoing" some of the installer and was a pain in the neck to fix). This update method fixes that by instructing the user to download and reupload their app and db files.